Added editTask test
